requirements.txt
flask>=2.3.0
werkzeug>=2.3.0
openai==0.27.8
python-dotenv==1.0.0
requests==2.28.2
pymongo==4.3.3
pyngrok==5.2.1
schedule==1.2.0
pytz==2023.3
openpyxl==3.1.2